update style, use styled instead of createStyles
